migration dan model

+++ database/migrations/2019_08_08_062226_create_gurus_table.php
@@ -16,7 +16,7 @@ class CreateGurusTable extends Migration
         Schema::create('gurus', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('guru_kode')->unique();
-            $table->unsignedInteger('kompetensi_id');
+            $table->unsignedBigInteger('kompetensi_id');
             $table->string('guru_NIP')->unique();
             $table->string('guru_nama');
             $table->text('guru_alamat');
